+++ sql_db_mapper_core/src/lib.rs
@@ -1,12 +1,17 @@
 //! Helper types and functions for auto-generateed psql database wrappers
 //!
-//! Provides the TryFromRow trait which converts from a tokio_postgres Row. Implementations are provided for common types
+//! Provides the [`TryFromRow`] trait which converts from a [`tokio_postgres::Row`]. Implementations are provided for common types
 //!
-//! Also contains Interval which represents a SQL Interval
+//! Also contains [`Interval`] which represents a SQL Interval
 //!
-//! Reexports tokio_postgres::Error as SqlError (the Result::Err of the return from TryFromRow::from_row) and tokio_postgres::row::Row
+//! Reexports [`tokio_postgres::Error`] as SqlError (the Result::Err of the return from [`TryFromRow::from_row`]) and [`tokio_postgres::Row`]
+//!
+//! [`tokio_postgres::Error`]: https://docs.rs/tokio-postgres/0.5.2/tokio_postgres/error/struct.Error.html
+//! [`tokio_postgres::Row`]: https://docs.rs/tokio-postgres/0.5.2/tokio_postgres/row/struct.Row.html
+//! [`TryFromRow::from_row`]: ./trait.TryFromRow.html#tymethod.from_row
+//! [`TryFromRow`]: ./trait.TryFromRow.html
+//! [`Interval`]: ./struct.Interval.html
 
-/// Contains and reexports types that are used by the TryFromRow trait
 pub use tokio_postgres::{
 	Error as SqlError,
 	row::Row,
@@ -29,11 +34,25 @@ use time::Duration;
 
 pub use sql_db_mapper_derive::*;
 
+
+/// Converts from a [`tokio_postgres::Row`]. Implementations are provided for common types
+///
+/// [`TryFromRow`]: ./trait.TryFromRow.html
+/// [`tokio_postgres::Row`]: https://docs.rs/tokio-postgres/0.5.2/tokio_postgres/row/struct.Row.html
 pub trait TryFromRow: Sized {
 	fn from_row(row : Row) -> Result<Self, SqlError>;
 }
+#[cfg(feature = "with_serde")]
+use serde::{
+	Serialize,
+	Deserialize,
+};
 
-/// Wrapper type around a time::Duration that implements ToSql and FromSql
+/// Wrapper type around a [`time::Duration`] that implements [`ToSql`] and [`FromSql`]
+///
+/// [`time::Duration`]: https://docs.rs/time/0.2.6/time/struct.Duration.html
+/// [`ToSql`]: https://docs.rs/postgres-types/0.1.0/postgres_types/trait.ToSql.html
+/// [`FromSql`]: https://docs.rs/postgres-types/0.1.0/postgres_types/trait.FromSql.html
 #[cfg_attr(feature = "with_serde", derive(Serialize, Deserialize))]
 #[derive(Debug, Clone)]
 pub struct Interval {